deps(composer): Update tomatophp/filament-types to v2.0 and add new dependencies
The change updates the version of tomatophp/filament-types from v1.0 to v2.0 in composer.json. In composer.lock, it not only reflects this version change but also adds several new dependencies like composer/pcre, ezyang/htmlpurifier, maatwebsite/excel, and more. This is likely to introduce new functionality or improvements related to the filament - types package and its related components in the application.